Fix very weird (silent!) overflow in Webassembly when storing u32 + u8 in i64-2... y u do dis?

+++ src/compiler.rs
@@ -553,7 +553,7 @@ impl Compiler {
                 let num_locals = self.symbols.leave_context() as u8;
 
                 // Create function object and store as constant
-                let obj = Object::function(pos_start_function as u32, num_locals);
+                let obj = Object::function(pos_start_function as u16, num_locals);
                 let idx = self.add_constant(obj);
                 self.add_instruction(OpCode::Const, &[idx]);
